cleaned tests
